fix: disable autocast for GatedIPA

+++ profold2/model/evoformer.py
@@ -169,7 +169,6 @@ class EvoformerBlock(nn.Module):
         with autocast(enabled=False):
           # to default float
           m, x, t = m.float(), x.float(), tuple(map(lambda x: x.float(), t))
-
           s = self.frame_attn(m[:, 0],
                               mask=msa_mask[:, 0],
                               pairwise_repr=x,
@@ -188,8 +187,11 @@ class EvoformerBlock(nn.Module):
     # pairwise attention and transition
     with profiler.record_function('pair_attn'):
       # with autocast(enabled=False):
-      x = self.pair_attn(x.float(), mask=mask, msa_repr=m.float(), msa_mask=msa_mask,
-               shard_size=shard_size)
+      x = self.pair_attn(x,
+                         mask=mask,
+                         msa_repr=m,
+                         msa_mask=msa_mask,
+                         shard_size=shard_size)
       x = self.pair_ff(x) + x
 
     return x, m, t, mask, msa_mask
@@ -203,7 +205,13 @@ class Evoformer(nn.Module):
     self.layers = nn.ModuleList(
         [EvoformerBlock(**kwargs) for _ in range(depth)])  # pylint: disable=missing-kwoa
 
-  def forward(self, x, m, frames=None, mask=None, msa_mask=None, shard_size=None):
+  def forward(self,
+              x,
+              m,
+              frames=None,
+              mask=None,
+              msa_mask=None,
+              shard_size=None):
     with profiler.record_function('evoformer'):
       inp = (x, m, frames, mask, msa_mask)
       x, m, t, *_ = checkpoint_sequential_nargs(self.layers,
